requirements.txt
discord.py>=2.3.0
pytz>=2024.1
apscheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
yt-dlp>=2024.1.0
PyNaCl>=1.5.0